pyproject.toml
[project]
name = "comfyui-umeairt-toolkit"
version = "1.0.0"
description = "A Pipeline-ready, Nodes 2.0 Ready, and Aesthetic Toolkit for ComfyUI."
readme = "README.md"
authors = [
    { name = "UmeAiRT Team", email = "contact@umeairt.com" }
]
license = { file = "LICENSE" }
dependencies = [
    "colorama",
    "opencv-python",
    "tqdm",
    "scipy",
    "psutil",
    "safetensors"
]
requires-python = ">=3.10"
keywords = ["comfyui", "stable-diffusion", "pipeline", "workflow", "ai"]

[project.optional-dependencies]
seedvr2 = [
    "diffusers>=0.33.1",
    "peft>=0.17.0",
    "rotary_embedding_torch>=0.5.3",
    "einops",
    "omegaconf>=2.3.0"
]
facedetailer = ["ultralytics>=8.0.0"]
gguf = ["gguf"]
all = [
    "diffusers>=0.33.1",
    "peft>=0.17.0",
    "rotary_embedding_torch>=0.5.3",
    "einops",
    "omegaconf>=2.3.0",
    "ultralytics>=8.0.0",
    "gguf"
]

[project.urls]
Homepage = "https://github.com/UmeAiRT/ComfyUI-UmeAiRT-Toolkit"
Repository = "https://github.com/UmeAiRT/ComfyUI-UmeAiRT-Toolkit.git"

[tool.comfy]
PublisherId = "UmeAiRT"
DisplayName = "UmeAiRT Toolkit"
Icon = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
import_mode = "importlib"
norecursedirs = ["modules", "vendor", "usdu_core", "facedetailer_core", "seedvr2_core", "web", "assets", ".git"]
collect_ignore_glob = ["*/__init__.py", "__init__.py"]

[tool.bandit]
# B101 = assert usage (fine in non-production code)
# B310 = urlopen audit (legitimate — we download models)
skips = ["B101", "B310"]
exclude_dirs = ["vendor", "usdu_core", "facedetailer_core", "seedvr2_core", "tests"]
